Update to check command executable

+++ update.sh
@@ -73,37 +73,55 @@ update() {
 
   msg_title "Start global update"
 
-  msg_heading "Update packages via apt-get"
-  msg_step "Update packages"
-  output_box sudo apt-get update
-  msg_step "Upgrade packages"
-  output_box sudo apt-get -y upgrade
+  if [[ -f "/usr/bin/apt-get" ]]; then
+    msg_heading "Update packages via apt-get"
+    msg_step "Update packages"
+    output_box sudo apt-get update
+    msg_step "Upgrade packages"
+    output_box sudo apt-get -y upgrade
+  fi
 
-  msg_heading "Update asdf-vm"
-  msg_step "Fetch from git"
-  output_box asdf update --head
+  if [[ -f "$HOME/.asdf" ]]; then
+      msg_heading "Update asdf-vm"
+    msg_step "Fetch from git"
+    output_box asdf update --head
+  fi
 
-  msg_heading "Update AWS Cli v2"
-  msg_step "Fetch from online"
-  curl -s "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" \
-    -o "awscliv2.zip"
-  mute unzip awscliv2.zip
-  msg_step "Update via script"
-  output_box sudo ./aws/install --update
-  msg_step "Clean up"
-  rm -rf ./aws
+  if [[ -f "$HOME/.aws" ]]; then
+    msg_heading "Update AWS Cli v2"
+    msg_step "Fetch from online"
+    curl -s "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" \
+      -o "awscliv2.zip"
+    mute unzip awscliv2.zip
+    msg_step "Update via script"
+    output_box sudo ./aws/install --update
+    msg_step "Clean up"
+    rm -rf ./aws
+  fi
 
-  msg_heading "Update via snap"
-  msg_step "Update snap packages"
-  output_box sudo snap refresh
+  if [[ -f "$HOME/snap" ]]; then
+    msg_heading "Update via snap"
+    msg_step "Update snap packages"
+    output_box sudo snap refresh
+  fi
 
-  msg_heading "Update via yarn"
-  msg_step "Update yarn packages"
-  output_box yarn global upgrade
+  if [[ -f "/usr/bin/yarn" ]]; then
+    msg_heading "Update via yarn"
+    msg_step "Update yarn packages"
+    output_box yarn global upgrade
+  fi
 
-  msg_heading "Update zinit, zinit plugins"
-  msg_step "Update zinit"
-  output_box zinit self-update
-  msg_step "Update zinit plugins"
-  output_box zinit update
+  if [[ -f "$HOME/.zinit" ]]; then
+    msg_heading "Update zinit, zinit plugins"
+    msg_step "Update zinit"
+    output_box zinit self-update
+    msg_step "Update zinit plugins"
+    output_box zinit update
+  fi
+
+  if [[ -f "$HOME/.config/ulauncher/user-themes/one-dark-ulauncher" ]]; then
+    msg_heading "Update ulauncher theme"
+    msg_step "Fetch from git"
+    output_box git -C ~/.config/ulauncher/user-themes/one-dark-ulauncher pull origin master
+  fi
 }
